requirements.txt
attrs==23.2.0
branca==0.7.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.9.6
folium==0.16.0
geopandas==0.14.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
selenium==4.18.1
shapely==2.0.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
xyzservices==2023.10.1
